reduced number of mysql queries. matchTimeout()
Rewrote payment scripts in declarewinner.php so that the sum of a user's
bets are combined before updating their points in user.sql.

Made a new function called matchTimeout() in functionlist.php since the
old match timeout script on betmain.php was limited to 10 entries and
also didn't refund points yet. I also found a better way to check the
age of the timestamp.